+++ Core/World/WorldStatic.cs
@@ -42,6 +42,7 @@ public static class WorldStatic
     public static bool FinalDoomTeleport;
     public static bool VanillaSectorSound;
     public static bool Udmf;
+    public static bool SectorFriction;
     public static EntityManager EntityManager = null!;
     public static WorldSoundManager SoundManager = null!;
     public static List<EntityFrame> Frames = null!;